Reorganize test package structure
* moving demos to `demo`
* legacy code to `legacy`
* backend-specific tests to respective backend sub-packages
