+++ src/test/java/sc/fiji/ome/zarr/pyramid/Pyramidal5DImageDataOpenTest.java
@@ -40,6 +40,8 @@ import org.junit.jupiter.api.Test;
 import org.scijava.Context;
 import org.scijava.prefs.PrefService;
 
+import sc.fiji.ome.zarr.pyramid.backend.n5.N5BackedPyramidal5DImageDataTest;
+import sc.fiji.ome.zarr.pyramid.backend.zarrjava.ZarrJavaBackedPyramidal5DImageDataTest;
 import sc.fiji.ome.zarr.pyramid.exceptions.NoMatchingResolutionException;
 import sc.fiji.ome.zarr.settings.ZarrOpeningSettings;
 import sc.fiji.ome.zarr.settings.ZarrReaderBackend;
