+++ src/test/java/sc/fiji/ome/zarr/pyramid/backend/zarrjava/ZarrJavaBackedPyramidal5DImageDataTest.java
@@ -26,30 +26,25 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package sc.fiji.ome.zarr.pyramid;
+package sc.fiji.ome.zarr.pyramid.backend.zarrjava;
 
-import java.nio.file.Paths;
+import java.net.URISyntaxException;
+import java.nio.file.Path;
 
-import net.imagej.ImageJ;
+import org.scijava.Context;
 
-import sc.fiji.ome.zarr.plugins.OpenInBDVCommand;
+import sc.fiji.ome.zarr.pyramid.Pyramidal5DImageDataImpl;
+import sc.fiji.ome.zarr.pyramid.Pyramidal5DImageDataTestBase;
+import sc.fiji.ome.zarr.util.ZarrTestUtils;
 
-public class Pyramidal5DImageDataDemo
+public class ZarrJavaBackedPyramidal5DImageDataTest implements Pyramidal5DImageDataTestBase
 {
-	public static void main( String[] args )
+	@Override
+	@SuppressWarnings( { "rawtypes", "unchecked" } )
+	public Pyramidal5DImageDataImpl< ?, ? > load( final String resource, final Context context, final Integer preferredWidth )
+			throws URISyntaxException
 	{
-		// final String multiscalePath = "https://uk1s3.embassy.ebi.ac.uk/idr/zarr/v0.4/idr0079A/idr0079_images.zarr/0";
-		final String multiscalePath = "/Users/hahmann/Data/idr0079_images.zarr/0"; // https://uk1s3.embassy.ebi.ac.uk/idr/zarr/v0.4/idr0079A/idr0079_images.zarr/0
-
-		// Show as imagePlus
-		final ImageJ imageJ = new ImageJ();
-		imageJ.ui().showUI();
-		final Pyramidal5DImageDataImpl< ?, ? > pyramidal5DImageData =
-				new Pyramidal5DImageDataImpl<>( imageJ.context(), Paths.get( multiscalePath ).toUri() );
-		PyramidalDataset< ? > pyramidalDataset = pyramidal5DImageData.asPyramidalDataset();
-		imageJ.ui().show( pyramidalDataset );
-
-		// Also show the displayed image in BDV
-		imageJ.command().run( OpenInBDVCommand.class, true );
+		Path path = ZarrTestUtils.resourcePath( resource );
+		return new Pyramidal5DImageDataImpl( context, new ZarrJavaPyramidBackend( path.toUri(), preferredWidth ) );
 	}
 }
