+++ src/test/java/sc/fiji/ome/zarr/pyramid/legacy/MultiscaleImage.java
@@ -6,13 +6,13 @@
  * %%
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
- * 
+ *
  * 1. Redistributions of source code must retain the above copyright notice,
  *    this list of conditions and the following disclaimer.
  * 2. Redistributions in binary form must reproduce the above copyright notice,
  *    this list of conditions and the following disclaimer in the documentation
  *    and/or other materials provided with the distribution.
- * 
+ *
  * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
  * AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
  * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
@@ -26,11 +26,11 @@
  * POSSIBILITY OF SUCH DAMAGE.
  * #L%
  */
-package sc.fiji.ome.zarr.pyramid;
+package sc.fiji.ome.zarr.pyramid.legacy;
 
 import sc.fiji.ome.zarr.pyramid.exceptions.NotAMultiscaleImageException;
 
-import static sc.fiji.ome.zarr.pyramid.Multiscales.MULTI_SCALE_KEY;
+import static sc.fiji.ome.zarr.pyramid.legacy.Multiscales.MULTI_SCALE_KEY;
 
 import net.imglib2.RandomAccessibleInterval;
 import net.imglib2.Volatile;
